// src/api/models/ConceptClassRetrieve.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConceptClassRetrieve = {
    concept_class_id: string;
    readonly created_at: string;
    readonly updated_at: string;
    concept_class_name?: string | null;
    concept_class_concept?: number | null;
};