// src/api/models/ConceptUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConceptUpdate = {
    readonly concept_id: number;
    concept_name?: string | null;
    concept_class?: string | null;
    concept_code?: string | null;
    valid_start_date?: string | null;
    valide_end_date?: string | null;
    domain?: string | null;
    vocabulary?: string | null;
};