// src/api/models/DomainRetrieve.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DomainRetrieve = {
  domain_id: string;
  readonly created_at: string;
  readonly updated_at: string;
  domain_name?: string | null;
  domain_concept?: number | null;
};